Altera as funcionalidades de busca, redirecionamento e adiciona um carousel nos cards
Atualiza a funcionalidade que busca as informações do filme para usar o serviço de API (services) em vez do .env

Refatora o redirecionamento para usar navigate com a rota do Profile e o ID do filme

Introduz a funcionalidade de Carousel para a exibição de cards
